rsx: Improve handling of aligned memory
